Add minor useful 'tools' to better understand html representation + add new container
The last one needs explanation. The new container is added for flexibility. Since the elements inside the about-me container needs to be at the center of the page, you could theorically put them on center but it could be difficult and slow process to do. So wrapping another container is a good alternative to center the elements, leaving the body unchanged